Instrument HTTP request end-to-end
This commit extends the current span to include the transfer of the
HTTP response body rather than ending it when the headers complete. This
provides a fuller, more accurate picture of the HTTP request.

This is implemented using a `TransformStream` on the HTTP response body
that acts as a pass-through and only calling `span.end()` once the
stream has completed.